+++ output.go
@@ -11,16 +11,16 @@ type Output struct {
 }
 
 type OutputData struct {
-	Concepts *OutputConcepts `json:"concepts,omitempty"`
+	Concepts []*OutputConcept `json:"concepts,omitempty"`
 	Image    *ImageData      `json:"image,omitempty"`
 	Metadata *interface{}    `json:"metadata,omitempty"`
 }
 
-type OutputConcepts []*OutputConcept
-
 type OutputConcept struct {
-	AppID *string `json:"app_id,omitempty"`
-	ID    string  `json:"id"`
-	Name  string  `json:"name,omitempty"`
-	Value float64 `json:"value,omitempty"`
+	AppID     *string `json:"app_id,omitempty"`
+	CreatedAt string  `json:"created_at,omitempty"`
+	ID        string  `json:"id"`
+	Name      string  `json:"name,omitempty"`
+	UpdatedAt string  `json:"updated_at,omitempty"`
+	Value     float64 `json:"value,omitempty"`
 }